Fix buffer release bug.
In the callback the buffer should be either released back to
the pool, or returned to the component, never both. (It can
be set so that the buffer release callback will retrieve the
buffer from the pool and return it to the component).

Fix it so that it is returned to the component if the port is
enabled, otherwise it is released.